[web-server] Integrate pulsar and web-server
This changeset is a proof-of-concept that integrates the web-server
and the pulsar framework. Offline, we set up a pulsar system
with 3 brokers, 3 zoo-keepers and 3 book-keepers.

Here, we receive messages from the client (using the message event)
and send it to the pulsar broker. We also simulate a consumer
using a separate thread that receives messages from the broker
and writes it to disk.

In this implementation, we have basically simulated the following
data movement: client -> server -> pulsar broker -> consumer.