[nlp-ers] Introduce Pulsar consumers.
This changeset introduces pulsar NLP consumers that subscribe
to topic suggest-topic. At this point in time, we simply route
messages back to the web-server throug a new topic: suggestions-topic.

In future changesets, this consumer will be enhanced to perform NLP
on user submitted text, query elastic search and return list of
suggestions.