[web-client] Receive real time suggestions.
This changeset is the first working prototype that returns
suggestions to client based on the entered text. We replace
the previous handshake messages between the client and server
with real messages and suggestions. We also add some response
time telemetry, and a sequence id which will be useful to process
messages if they arrive out-of-order.

Other minor changes include replacing hard-coded URL with
environment values and including logging API for cleaner logs.