[batch-pipeline] Index S3 data into ES using Spark
In this changeset, we use Apache Spark to read
data from S3 and index in to AWS Elastic Search (ES).

Offline, we first setup a spark cluster and an elastic
search cluster. The elastic search cluster URL is
exposed as an environment variable and we use spark
to project (get a subset of columns) and index into ES.