[search] Add pulsar curator.
We add a new curator consumer that reads from the
curate-topic channel and writes currated suggestion
to the suggestions-channel.

The curator uses the search results from ES to query
Postgres and get the count of answers for each question.
The results are scored based on the product of ES score
and answer count and returned back to the user. More
sophisticated curation tips will be implemented later.

The search-es consumers only queries ES and send
results to the curator.