[web-client] Handle high input message rate.
Until now, we are primarily rate limiting messages to 1 per second.
As we started increasing the message rate, we noticed that the main
loop of the web-client needs to block until all responses are returned.
We add socket-io.wait() to address this issue. We also use socketio's
sleep method instead of python's sleep method during the rate limiting
process.