[search] Get ES index from packet (if present).
In order to optimize ES performance, we want to experiment with
different indexing schemes. For example, instead of one common
index for the entire dataset, we could have an index per stackexchange
subdomain (or site) as they call it. In order to experiment with
different indexing options, pass the name of the index to the
consumer in the body of the message. If no index is present, use
the default index.