[batch-pipeline] Add custom indexing options.
In order to study the performance impact of index layout
on ES queries, we need to have different indices. In this
changeset, we add an option to add custom indices based
on the post subdomain. If post subdomain is stackoverflow,
we create an index called stackoverflow. Otherwise, we
create an index called other.